requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
itsdangerous==2.1.2
python-dotenv==1.0.1
supabase==2.4.0
httpx==0.25.2
gotrue==2.9.1
postgrest==0.16.11
realtime==1.0.6
redis==5.0.1
requests==2.31.0
PyMySQL==1.1.0
python-magic==0.4.27
sendgrid==6.11.0
bcrypt==4.1.2
passlib==1.7.4
python-dateutil==2.9.0
uuid==1.30
websockets==12.0
